backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
azure-identity==1.15.0
azure-cosmos==4.5.1
azure-search-documents==11.4.0
azure-ai-contentsafety==1.0.0
azure-ai-projects==1.0.0b1
semantic-kernel==0.9.0b1
openai==1.12.0
aiohttp==3.9.3
redis==5.0.1
